# libs/contracts/rest/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "CreateTopicRequestDto": {
        "properties": {},
        "type": "object"
      },
...
      "RecordSummaryFeedbackRequestDto": {
        "properties": {
          "category": {
            "enum": [
              "wrong_fact",
              "missing_source",
              "bad_citation",
              "low_relevance",
              "too_verbose",
              "too_terse",
              "source_request",
              "ux_confusing",
              "other"
            ],
            "type": "string"
          },
          "citationId": {
            "type": "string"
          },
          "comment": {
            "maxLength": 2000,
            "type": "string"
          },
          "rating": {
            "maximum": 5,
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "category",
          "rating"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {},
    "description": "Backend/API-first social monitoring MVP.",
    "title": "Social Monitor API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/summaries/{summaryId}/feedback": {
      "get": {
        "operationId": "SummaryFeedbackController_list",
        "parameters": [
          {
            "in": "path",
            "name": "summaryId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "description": "Optional Bearer API key. Requires read:summaries. If supplied, x-workspace-role is not required.",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated workspace roles. Summary feedback reads allow owner, admin, member or viewer. Required when Authorization bearer API key is not supplied.",
            "in": "header",
            "name": "x-workspace-role",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List classified feedback for one summary with cursor pagination.",
        "tags": [
          "summaries"
        ]
      },
      "post": {
        "operationId": "SummaryFeedbackController_create",
        "parameters": [
          {
            "in": "path",
            "name": "summaryId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional actor id. API-key requests fall back to the API key id when omitted.",
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional Bearer API key. Requires write:summaries. If supplied, x-workspace-role is not required.",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated workspace roles. Summary feedback allows owner, admin, member or viewer. Required when Authorization bearer API key is not supplied.",
            "in": "header",
            "name": "x-workspace-role",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordSummaryFeedbackRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Record classified feedback for a summary without mutating the artifact.",
        "tags": [
          "summaries"
        ]
      }
    },
...
    "/topics": {
      "get": {
        "operationId": "TopicController_list",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "description": "Optional Bearer API key. Requires read:topics. If supplied, x-workspace-role is not required.",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated workspace roles. Topic reads allow owner, admin, member or viewer. Required when Authorization bearer API key is not supplied.",
            "in": "header",
            "name": "x-workspace-role",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List topics inside the current tenant/workspace.",
        "tags": [
          "topics"
        ]
      },
      "post": {
        "operationId": "TopicController_create",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional Bearer API key. Requires write:topics. If supplied, x-workspace-role is not required.",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated workspace roles. Topic creation requires owner or admin. Required when Authorization bearer API key is not supplied.",
            "in": "header",
            "name": "x-workspace-role",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTopicRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a topic inside the current tenant/workspace.",
        "tags": [
          "topics"
        ]
      }
    },
...
  },
  "servers": [],
...
}